Merged PR 6663: Toggle Restaurat Visibility
Added a VisibilityDialog component that displays a modal when the visibility icon is clicked.

The modal asks for confirmation before updating the restaurant's visibility:

If the restaurant is currently visible, the modal asks: “Do you want to hide this restaurant from users?”

If the restaurant is currently hidden, the modal asks: “Do you want to make this restaurant visible to users?”

Clicking Yes updates the visibility through the API, while No cancels the action.

Added responsive styling and rounded corners for better UI/UX.

Toast notifications are shown after successfully updating visibility to provide feedback to the user.

![image.png](https://dev.azure.com/ikentechnology/c8ab89bb-3867-421d-9adc-13cea45e64c7/_apis/git/repositories/a4357018-059a-4270-afaf-53d791ca4d9b/pullRequests/6663/attachments/image.png)
![image (4).png](https://dev.azure.com/ikentechnology/c8ab89bb-3867-421d-9adc-13cea45e64c7/_apis/git/repositories/a4357018-059a-4270-afaf-53d791ca4d9b/pullRequests/6663/attachments/image%20%284%29.png)
![image (5).png](https://dev.azure.com/ikentechnology/c8ab89bb-3867-421d-9adc-13cea45e64c7/_apis/git/repositories/a4357018-059a-4270-afaf-53d791ca4d9b/pullRequests/6663/attachments/image%20%285%29.png)
![image (6).png](https://dev.azure.com/ikentechnology/c8ab89bb-3867-421d-9adc-13cea45e64c7/_apis/git/repositories/a4357018-059a-4270-afaf-53d791ca4d9b/pullRequests/6663/attachments/image%20%286%29.png)

Related work items: #9330